--- src/utils/log.ts
@@ -1,5 +1,0 @@
-module.exports = function debugLog() {
-  if (process.env.NODE_ENV === 'development') {
-    return console.log(Array.prototype.slice.call(arguments))
-  }
-}
